backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pypdf==4.3.1
openai==1.51.0
deep-translator==1.11.4
python-dotenv==1.0.1
